telux: Allow partial reconfiguration, test app improvements
- Allow partial reconfiguration of sensor
- Sensor sample apps to consider command line arguments
- Sensor range data-type corrected as integer since it is
an index in the /etc/sensors.conf and not the value itself
- Menu options are named better
- Format output for clarity
- Improved sanity check for inputs via stdin
- Remove Cleanup_Reinit from menu options
- Limit sampling rate fixed point notation to 2 digits
- List created sensors to indicate the current sensor configuration
- Clean-up of SensorFeature menu to disable all enabled features
- Retain menu objects throughout the lifetime of the app

CRs-Fixed: 2927326, 2928553, 2934359, 2935061, 2927319, 2948191

+++ include/telux/sensor/SensorDefines.hpp
@@ -107,7 +107,7 @@ struct SensorInfo {
      * For gyroscopes, this is the number of degrees per second (dps) in either direction (+/-)
      * along each axis
      */
-    float range;
+    int range;
 };
 
 /**
@@ -198,8 +198,10 @@ struct SensorConfiguration {
      *
      * Further, this bitset should be set by the user to indicate the valid fields while configuring
      * the sensor using @ref telux::sensor::ISensor::configure.
-     * For continuous stream of data from a sensor, SAMPLING_RATE and BATCH_COUNT from
-     * @ref SensorConfigParams should be set
+     * For continuous stream of data from a sensor, the validity of SAMPLING_RATE and BATCH_COUNT
+     * from @ref SensorConfigParams should be considered. If the sensor had been already configured
+     * with both sampling rate and batch count, it is possible to reconfigure the sensor partially
+     * with just one of these attributes and setting the required validity flag.
      */
     SensorConfigMask validityMask;
